stop truncating the edit parameters pointer on the way to the window
Opening a map killed the editor. The stack is CDWGDBBrowser::Open ->
CEditorContainer::CreateNewEditor -> CMapInfoState::Enter ->
CFieldState::HandleCommand -> CFieldWindow::SetEditParameters, faulting
fourteen bytes into a function whose first statement writes a member.

That looked like a null this and was not. rcx held a healthy heap
pointer; rdx, the SEditParameters reference, held 0x00000000843c0210,
which is 0x00000212843c0210 with the top half cut off.

SetGetEditParameters took its pointer as uint32_t. Every caller is
SetEditParameters or GetEditParameters in ED_B2_M1/EditorMethods.h,
which pass reinterpret_cast< uintptr_t >( &params ), and the far end
casts the value straight back to a pointer and dereferences it. On x64
the cast to uintptr_t is a no-op and the narrowing to uint32_t is the
whole bug. ICommandHandler::HandleCommand already takes uintptr_t, so
this parameter was the only place in the chain where the pointer did not
fit.

Verified by repeating the sequence that produced the crash: post
ID_MAIN_OPEN, walk the object tree down to
Editor\Backgrounds\Summer\MapInfo.xdb, select it, post IDOK. Before, the
editor was gone within seconds and crashpad wrote a dump. After, it
opens the map: the frame title becomes "Blitzkrieg 2 Editor -
[Editor\Backgrounds\Summer\...", an MDI document exists, and the MapInfo
Tools toolbar comes up 223x26 where it had been 0x0 -- it belongs to an
editor module whose document was never active before, which is why the
seven hidden toolbars looked like a toolbar problem and were not.

This is the first map ever opened in this port. It is worth saying that
the parts it exercised were already working: the Open dialog's tree
filled with 278 MapInfo records through the forwarded SysTreeView32, and
TVM_SELECTITEM reached OnSelChanged through ON_NOTIFY_REFLECT well
enough for the dialog to update its Current Selection label and enable
OK. The tree was never the problem.

+++ Versions/Temporary/Engine/Sources/MapEditorLib/CommonEditorMethods.h
@@ -4,7 +4,18 @@
 
 #include "MapEditorLib_export.h"
 
-MAPEDITORLIB_EXPORT bool SetGetEditParameters( uint32_t pEditParameters, unsigned nCommandHandlerType, int nCmdID );
+//! Route an SEditParameters pointer to a window through the command handler.
+//!
+//! uintptr_t and not uint32_t. Every caller is
+//! SetEditParameters/GetEditParameters in ED_B2_M1/EditorMethods.h, which pass
+//! reinterpret_cast<uintptr_t>( &params ), and the receiving end casts the
+//! value straight back to a pointer and dereferences it. A uint32_t here threw
+//! away the top half on x64: opening a map faulted in
+//! CFieldWindow::SetEditParameters reading 0x00000000843c0210, which is
+//! 0x00000212843c0210 with the high word gone. ICommandHandler::HandleCommand
+//! already takes uintptr_t, so this parameter was the only narrowing in the
+//! chain.
+MAPEDITORLIB_EXPORT bool SetGetEditParameters( uintptr_t pEditParameters, unsigned nCommandHandlerType, int nCmdID );
 MAPEDITORLIB_EXPORT void CreateRefKey( std::string *pszKey, const struct SPropertyDesc *pPropertyDesc );
